v1.5 refactor(TextTools): update .gitignore and fix minor bugs
Update .gitignore to exclude unnecessary files and folders such as .idea,
TextTools-code/dist, and TextTools-code/Interface.build. Additionally, fix
a bug in the makeDocx function where the Filename was not correctly
assigned, and address minor code cleanup for readability and maintenance.

The update also includes changing the version in Interface.py from v1.4 to
v1.5 and modifying the corresponding softName to reflect the version
update.